+++ app/helpers/exporters/helpers.py
@@ -1,0 +1,9 @@
+def format_timedelta(td):
+    hours, remainder = divmod(td.total_seconds(), 3600)
+    minutes, seconds = divmod(remainder, 60)
+    hours, minutes, seconds = int(hours), int(minutes), int(seconds)
+    if hours < 10:
+        hours = '0%s' % int(hours)
+    if minutes < 10:
+        minutes = '0%s' % minutes
+    return '%s:%s' % (hours, minutes)
